+++ modules/gameboard/src/GamePlayStateMachine copy.js
@@ -1,11 +1,10 @@
 import { forwardRef, useContext, useEffect, useRef, useState } from 'react';
 import StateMachine from 'javascript-state-machine';
+import {GamePlayContext} from './GamePlayContext';
 import * as TEXT from './loadText';
-import {PlayContext} from './PlayContext';
 import AudioPlayer from './AudioPlayer';
 
-const PlayStateMachine = forwardRef(({ ballRef }, ref) => {
-
+const GamePlayStateMachine = forwardRef(({ ballRef }, ref) => {
     const {
         setCountdown,
         setTopPaddleState,
@@ -17,13 +16,13 @@ const PlayStateMachine = forwardRef(({ ballRef }, ref) => {
         setIncludeCountDown,
         // setIsTopPaddleReset,
         // setIsBottomPaddleReset,
-        // setIsBallReset,
+        setIsBallReset,
         setPrevLevel,
         setForcePaddleRerender,
         setForceBallRerender,
         setResetBall,
         setIsDontPlay,
-    } = useContext(PlayContext);  
+    } = useContext(GamePlayContext);  
 
     const audioPlayerRef = useRef(null);
     const audioVolume = 0.5;
@@ -43,65 +42,77 @@ const PlayStateMachine = forwardRef(({ ballRef }, ref) => {
             init: 'gameStarted',
             transitions: [
                 { name: 'startGame', from: ['gameStarted'], to: 'idle' },
-                { name: 'resetLevel', from: ['gameReset'], to: 'levelReset' },
-                { name: 'startLevel', from: 'levelReset', to: 'paddleReset'},
-                { name: 'startCountdown', from: 'gameReset', to: 'countdown'},
-                { name: 'beginPlay', from: ['countdown', 'gameReset'], to: 'play'},
+                // { name: 'resetLevel', from: ['gameReset'], to: 'levelReset' },
+                { name: 'resetPaddle', from: ['idle', 'topGoal', 'bottomGoal'], to: 'paddleReset' },
+                { name: 'resetBall', from: 'paddleReset', to: 'ballReset' },
+                // { name: 'resetLevel', from: ['ballReset'], to: 'levelReset' },
+
+                // { name: 'resetLevel', from: ['ballReset'], to: 'countdown' },
+                // { name: 'resetPlay', from: ['ballReset'], to: 'startPlay' },
+
+                // { name: 'startLevel', from: 'levelReset', to: 'paddleReset'},
+                // { name: 'startCountdown', from: 'gameReset', to: 'countdown'},
+                // { name: 'startCountdown', from: 'levelReset', to: 'countdown'},
+                { name: 'startCountdown', from: 'ballReset', to: 'countdown'},
+                // { name: 'beginPlay', from: ['countdown', 'gameReset'], to: 'play'},
+
+                // { name: 'startPlay', from: ['countdown', 'levelReset'], to: 'play'},
+                { name: 'startPlay', from: ['countdown', 'ballReset'], to: 'play'},
+
                 { name: 'topGoalScored', from: 'play', to: 'topGoal'},
                 { name: 'bottomGoalScored', from: 'play', to: 'bottomGoal'},                
                 { name: 'endLevel', from: ['topGoal', 'bottomGoal', 'play', 'paddleReset'], to: 'idle'},
 
-                { name: 'resetPaddle', from: ['idle', 'topGoal', 'bottomGoal'], to: 'paddleReset' },
-                { name: 'resetBall', from: 'paddleReset', to: 'ballReset' },
-                { name: 'resetGame', from: 'ballReset', to: 'gameReset' },
+                // { name: 'resetGame', from: 'ballReset', to: 'gameReset' },
 
-                { name: 'endGame', from: ['gameReset'], to: 'gameFinished' },
+                // { name: 'endGame', from: ['gameReset'], to: 'gameFinished' },
+                { name: 'endGame', from: ['levelReset'], to: 'gameFinished' },
                 { name: 'returnToIdle', from: ['gameFinished'], to: 'idle' },
             ],
             methods: {
                 // Method called when entering the 'gameStarted' state
                 onEnterGameStarted: () => {
-                    console.log('PlayStateMachine Entering gameStarted state');
+                    console.log('GamePlayStateMachine Entering gameStarted state');
                     setPrevLevel(0);
                     setTopScore(0);
                     setBottomScore(0);
                 },
-                // Method called when leaving the 'gameStarted' state
+                // Method called when Leave the 'gameStarted' state
                 onLeaveGameStarted: () => {
-                    console.log('PlayStateMachine Leaving gameStarted state');
+                    console.log('GamePlayStateMachine Leave gameStarted state');
                 },                 
                 // Method called when entering the 'idle' state
                 onEnterIdle: () => {
-                    console.log('PlayStateMachine Entering idle state');
+                    console.log('GamePlayStateMachine Entering idle state');
                     setIsDontPlay(true);
                     setIncludeCountDown(false);
                     setCountdown(TEXT.countdown_get_ready);
                 },
-                // Method called when leaving the 'idle' state
+                // Method called when Leave the 'idle' state
                 onLeaveIdle: () => {
-                    console.log('PlayStateMachine Leaving idle state');
+                    console.log('GamePlayStateMachine Leave idle state');
                 },    
                 // Method called when entering the 'levelReset' state
                 onEnterLevelReset: () => {
-                    console.log('PlayStateMachine Entering levelReset state');
+                    console.log('GamePlayStateMachine Entering levelReset state');
                     setIsDontPlay(false);
                     setIsCountdownComplete(false);
                     setIncludeCountDown(true);
                     setCountdown(TEXT.blank);
                 },
-                // Method called when leaving the 'levelReset' state
+                // Method called when Leave the 'levelReset' state
                 onLeaveLevelReset: () => {
-                    console.log('PlayStateMachine Leaving levelReset state');
+                    console.log('GamePlayStateMachine Leave levelReset state');
                 },  
                 // Method called when entering the 'paddleReset' state
                 onEnterPaddleReset: () => {
-                    console.log('PlayStateMachine Entering paddleReset state');
+                    console.log('GamePlayStateMachine Entering paddleReset state');
                     setResetPaddles(true);
                     setForcePaddleRerender(prevState => !prevState);
                 },
-                // Method called when leaving the 'paddleReset' state
+                // Method called when Leave the 'paddleReset' state
                 onLeavePaddleReset: () => {
-                    console.log('PlayStateMachine Leaving paddleReset state');
+                    console.log('GamePlayStateMachine Leave paddleReset state');
                     setTopPaddleState("play");
                     setBottomPaddleState("play");
                     // setIsTopPaddleReset(false);
@@ -111,59 +122,59 @@ const PlayStateMachine = forwardRef(({ ballRef }, ref) => {
                 },
                 // Method called when entering the 'ballReset' state
                 onEnterBallReset: () => {
-                    console.log('PlayStateMachine Entering ballReset state');
+                    console.log('GamePlayStateMachine Entering ballReset state');
                     setResetBall(true);
                     setForceBallRerender(prevState => !prevState);
                 },
-                // Method called when leaving the 'ballReset' state
+                // Method called when Leave the 'ballReset' state
                 onLeaveBallReset: () => {
-                    console.log('PlayStateMachine Leaving ballReset state');
+                    console.log('GamePlayStateMachine Leave ballReset state');
                     setIsBallReset(false);
                     setResetBall(false);
                     setForceBallRerender(prevState => !prevState);
                 },
-                // Method called when entering the 'reset' state
-                onEnterGameReset: () => {
-                    console.log('PlayStateMachine Entering gameReset state');
-                },
-                // Method called when leaving the 'reset' state
-                onLeaveGameReset: () => {
-                    console.log('PlayStateMachine Leaving gameReset state');
-                },
+                // // Method called when entering the 'reset' state
+                // onEnterGameReset: () => {
+                //     console.log('GamePlayStateMachine Entering gameReset state');
+                // },
+                // // Method called when Leave the 'reset' state
+                // onLeaveGameReset: () => {
+                //     console.log('GamePlayStateMachine Leave gameReset state');
+                // },
                 // Method called when entering the 'countdown' state
                 onEnterCountdown: () => {
-                    console.log('PlayStateMachine Entering countdown state');
+                    console.log('GamePlayStateMachine Entering countdown state');
                     setCountdown(TEXT.countdown_three);
                     setTimeout(() => {
-                    setCountdown(TEXT.countdown_two);
-                    setTimeout(() => {
-                        setCountdown(TEXT.countdown_one);
+                        setCountdown(TEXT.countdown_two);
                         setTimeout(() => {
-                        setCountdown(TEXT.countdown_go);
-                        setTimeout(() => {
-                            setCountdown(TEXT.blank);
-                            setIsCountdownComplete(true);     
-                        }, delay); 
-                        }, delay); 
-                    }, delay);
+                            setCountdown(TEXT.countdown_one);
+                            setTimeout(() => {
+                                setCountdown(TEXT.countdown_go);
+                                setTimeout(() => {
+                                    setCountdown(TEXT.blank);
+                                    setIsCountdownComplete(true);     
+                                }, delay); 
+                            }, delay); 
+                        }, delay);
                     }, delay);
                 },
-                // Method called when leaving the 'countdown' state
+                // Method called when Leave the 'countdown' state
                 onLeaveCountdown: () => {
-                    console.log('PlayStateMachine Leaving countdown state');
+                    console.log('GamePlayStateMachine Leave countdown state');
                 },
                 // Method called when entering the 'play' state
                 onEnterPlay: () => {
-                    console.log('PlayStateMachine Entering play state');
+                    console.log('GamePlayStateMachine Entering play state');
                     ballRef.current.setLinvel({ x: 0, y: speed, z: 0 }, true);
                 },
-                // Method called when leaving the 'play' state
+                // Method called when Leave the 'play' state
                 onLeavePlay: () => {
-                    console.log('PlayStateMachine Leaving play state');
+                    console.log('GamePlayStateMachine Leave play state');
                 },
                 // Method called when entering the 'topGoal' state
                 onEnterTopGoal: () => {
-                    console.log('PlayStateMachine Entering topGoal state');
+                    console.log('GamePlayStateMachine Entering topGoal state');
                     if (audioPlayerRef.current) {
                         audioPlayerRef.current.play('pointScore').catch((error) => {
                             console.error('Error playing audio:', error);
@@ -173,13 +184,13 @@ const PlayStateMachine = forwardRef(({ ballRef }, ref) => {
           
                     setIncludeCountDown(false);
                 },
-                // Method called when leaving the 'topGoal' state
+                // Method called when Leave the 'topGoal' state
                 onLeaveTopGoal: () => {
-                    console.log('PlayStateMachine Leaving topGoal state');
+                    console.log('GamePlayStateMachine Leave topGoal state');
                 },
                 // Method called when entering the 'bottomGoal' state
                 onEnterBottomGoal: () => {
-                    console.log('PlayStateMachine Entering bottomGoal state');
+                    console.log('GamePlayStateMachine Entering bottomGoal state');
                     if (audioPlayerRef.current) {
                         audioPlayerRef.current.play('pointLose').catch((error) => {
                             console.error('Error playing audio:', error);
@@ -189,20 +200,20 @@ const PlayStateMachine = forwardRef(({ ballRef }, ref) => {
             
                     setIncludeCountDown(false);
                 },
-                // Method called when leaving the 'bottomGoal' state
+                // Method called when Leave the 'bottomGoal' state
                 onLeaveBottomGoal: () => {
-                    console.log('PlayStateMachine Leaving bottomGoal state');
+                    console.log('GamePlayStateMachine Leave bottomGoal state');
                 },
                 // Method called when entering the 'gameFinished' state
                 onEnterGameFinished: () => {
-                    console.log('PlayStateMachine Entering gameFinished state');
+                    console.log('GamePlayStateMachine Entering gameFinished state');
                 },
-                // Method called when leaving the 'gameFinished' state
+                // Method called when Leave the 'gameFinished' state
                 onLeaveGameFinished: () => {
-                    console.log('PlayStateMachine Leaving gameFinished state');
+                    console.log('GamePlayStateMachine Leave gameFinished state');
                 },
                 onInvalidTransition: function(transition, from, to) {
-                    console.log("PlayStateMachine transition '%s' not allowed from state '%s'", transition, from);
+                    console.log("GamePlayStateMachine transition '%s' not allowed from state '%s'", transition, from);
                 },
             },
         });
@@ -215,4 +226,4 @@ const PlayStateMachine = forwardRef(({ ballRef }, ref) => {
     return null;
 });
 
-export default PlayStateMachine;
+export default GamePlayStateMachine;
