+++ modules/gameboard/src/MainScene.js
@@ -2,10 +2,10 @@ import React, {Suspense, useEffect, useContext} from 'react'
 import { Canvas, useThree } from "@react-three/fiber";
 import { Physics } from '@react-three/rapier';
 import {GizmoHelper, GizmoViewcube, GizmoViewport, GridHelper} from "@react-three/drei"; // can be commented in for debugging
-import Game from './Game';
-import GameHud from './GameHud';
-import PlayerInstructionsHud from './PlayerInstructionHud';
-import { GameContext } from './GameContext';
+import GamePlay from './GamePlay';
+import GamePlayHud from './GamePlayHud';
+import GameInstructionsHud from './GameInstructionHud';
+import { GameSessionContext } from './GameSessionContext';
 
 function CameraController() {
   const { camera } = useThree();
@@ -24,8 +24,8 @@ function CameraController() {
 function MainScene() {
 
   const {
-    isGame,
-  } = useContext(GameContext);
+    isGamePlayComplete,
+  } = useContext(GameSessionContext);
 
   return (
     <Canvas mode="concurrent">
@@ -33,15 +33,15 @@ function MainScene() {
       <CameraController />
       <Suspense fallback={null} >
         <Physics gravity={[0, 0, 0]} >
-        {isGame ? (
-          <group> 
-            <Game/>
-            <GameHud/>
-          </group>
-        ) : (
+        {isGamePlayComplete ? (
           <group>
-            <PlayerInstructionsHud/>
+            <GameInstructionsHud/>
           </group>  
+        ) : (
+          <group> 
+            <GamePlay/>
+            <GamePlayHud/>
+          </group>
         )}
         </Physics>   
       </Suspense>
