+++ modules/gameboard/src/imgs/gameInstructions/GameInstructions-MoveLeft.png
