+++ modules/gameboard/src/imgs/gameInstructions/GameInstructions-MoveLeftRight.png
