+++ modules/gameboard/src/imgs/gameInstructions/GameInstructions-MoveRight.png
